requirements.txt

numpy
pandas
matplotlib
tensorflow-gpu
